pyproject.toml
[project]
name = "pdf-ai-rename"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "loguru>=0.7.2",
    "openai>=1.52.2",
    "pypdf2>=3.0.1",
    "tiktoken>=0.8.0",
]
